feat:booking

+++ src/main/java/com/swd/ccp/repositories/CustomerRepo.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface CustomerRepo extends JpaRepository<Customer, Integer> {
-    Optional<Customer> findById(Long customerId);
+    Optional<Customer> findById(Integer customerId);
     Optional<Customer> findByAccount_Email(String email);
 }
